Improve retention readability
