+++ crates/zakura-header-chain/src/transition/planner/retention.rs
@@ -36,8 +36,8 @@ pub(super) struct RetentionWork {
 /// Enforce deterministic retention while protecting selected, verified, and context paths.
 pub(super) fn enforce_retention<G: HeaderGraphEdit>(
     store: &mut G,
-    header_best: Frontier,
-    verified_best: Frontier,
+    selected_header_tip: Frontier,
+    verified_header_tip: Frontier,
     validation_context_references: impl IntoIterator<Item = block::Hash>,
     limits: EngineLimits,
 ) -> Result<RetentionPlan, GraphError> {
@@ -49,24 +49,40 @@ pub(super) fn enforce_retention<G: HeaderGraphEdit>(
     }
 
     let mut plan = RetentionPlan::default();
-    let mut protected = HashSet::new();
+    let mut protected_header_hashes = HashSet::new();
     plan.work.graph_workspaces = 1;
-    plan.work.protected_path_visits += protect_path(store, header_best.hash, &mut protected)?;
-    plan.work.protected_path_visits += protect_path(store, verified_best.hash, &mut protected)?;
+    plan.work.protected_path_visits += add_protected_header_path(
+        store,
+        selected_header_tip.hash,
+        &mut protected_header_hashes,
+    )?;
+    plan.work.protected_path_visits += add_protected_header_path(
+        store,
+        verified_header_tip.hash,
+        &mut protected_header_hashes,
+    )?;
     for reference in validation_context_references {
-        if !protected.contains(&reference) && store.view_header_node(reference).is_some() {
-            plan.work.protected_path_visits += protect_path(store, reference, &mut protected)?;
+        if !protected_header_hashes.contains(&reference)
+            && store.view_header_node(reference).is_some()
+        {
+            plan.work.protected_path_visits +=
+                add_protected_header_path(store, reference, &mut protected_header_hashes)?;
         }
     }
 
-    if protected.len().saturating_sub(1) > limits.max_non_finalized_nodes.get() {
-        return Ok(stalled(plan));
+    if protected_header_hashes.len().saturating_sub(1) > limits.max_non_finalized_nodes.get() {
+        return Ok(resource_stalled_plan(plan));
     }
 
     let mut candidates = RetentionCandidates::build(store)?;
-    plan.work.candidate_nodes_scanned = candidates.scanned;
+    plan.work.candidate_nodes_scanned = candidates.header_nodes_scanned;
     plan.work.graph_workspaces = plan.work.graph_workspaces.saturating_add(1);
-    evict_permanently_ineligible(store, &protected, &mut candidates, &mut plan.work)?;
+    evict_permanently_ineligible(
+        store,
+        &protected_header_hashes,
+        &mut candidates,
+        &mut plan.work,
+    )?;
     if store.view_eligible_header_tip_count() <= limits.max_candidate_tips.get()
         && store.view_header_node_count().saturating_sub(1) <= limits.max_non_finalized_nodes.get()
     {
@@ -75,14 +91,19 @@ pub(super) fn enforce_retention<G: HeaderGraphEdit>(
 
     if store.view_eligible_header_tip_count() > limits.max_candidate_tips.get() {
         while store.view_eligible_header_tip_count() > limits.max_candidate_tips.get() {
-            let Some(score) = candidates.eligible_tips.pop_first() else {
-                return Ok(stalled(plan));
+            let Some(score) = candidates.eligible_header_tips.pop_first() else {
+                return Ok(resource_stalled_plan(plan));
             };
-            if protected.contains(&score.tip_hash) {
+            if protected_header_hashes.contains(&score.tip_hash) {
                 continue;
             }
             if store.view_header_node(score.tip_hash).is_some() {
-                evict_tip_branch(store, score.tip_hash, &protected, &mut plan.work)?;
+                evict_tip_branch(
+                    store,
+                    score.tip_hash,
+                    &protected_header_hashes,
+                    &mut plan.work,
+                )?;
             }
         }
     }
@@ -91,15 +112,15 @@ pub(super) fn enforce_retention<G: HeaderGraphEdit>(
         while store.view_header_node_count().saturating_sub(1)
             > limits.max_non_finalized_nodes.get()
         {
-            let Some(score) = candidates.leaves.pop_first() else {
-                return Ok(stalled(plan));
+            let Some(score) = candidates.header_leaves.pop_first() else {
+                return Ok(resource_stalled_plan(plan));
             };
             let hash = score.tip_hash;
-            if protected.contains(&hash) {
+            if protected_header_hashes.contains(&hash) {
                 continue;
             }
             if store.view_header_node(hash).is_some() {
-                evict_tip_branch(store, hash, &protected, &mut plan.work)?;
+                evict_tip_branch(store, hash, &protected_header_hashes, &mut plan.work)?;
             }
         }
     }
@@ -107,38 +128,38 @@ pub(super) fn enforce_retention<G: HeaderGraphEdit>(
     Ok(plan)
 }
 
-fn stalled(mut plan: RetentionPlan) -> RetentionPlan {
+fn resource_stalled_plan(mut plan: RetentionPlan) -> RetentionPlan {
     plan.admission_refused = true;
     plan.resource_stalled = true;
     plan
 }
 
 struct RetentionCandidates {
-    permanent_roots: BTreeSet<(block::Height, [u8; 32])>,
-    eligible_tips: BTreeSet<crate::ChainScore>,
-    leaves: BTreeSet<crate::ChainScore>,
-    scanned: usize,
+    permanently_ineligible_roots: BTreeSet<(block::Height, [u8; 32])>,
+    eligible_header_tips: BTreeSet<crate::ChainScore>,
+    header_leaves: BTreeSet<crate::ChainScore>,
+    header_nodes_scanned: usize,
 }
 
 impl RetentionCandidates {
     fn build<G: HeaderGraphView>(store: &G) -> Result<Self, GraphError> {
-        let mut permanent_roots = BTreeSet::new();
-        let mut eligible_tips = BTreeSet::new();
-        let mut leaves = BTreeSet::new();
-        let mut scanned = 0usize;
+        let mut permanently_ineligible_roots = BTreeSet::new();
+        let mut eligible_header_tips = BTreeSet::new();
+        let mut header_leaves = BTreeSet::new();
+        let mut header_nodes_scanned = 0usize;
         let mut error = None;
         store.visit_header_nodes(&mut |node| {
             if error.is_some() {
                 return;
             }
-            scanned = scanned.saturating_add(1);
+            header_nodes_scanned = header_nodes_scanned.saturating_add(1);
             if node.eligibility.has_permanent_reason()
                 || matches!(
                     node.body_validation_state,
                     BodyValidationState::ConsensusInvalid { .. }
                 )
             {
-                permanent_roots.insert((node.height, node.hash.0));
+                permanently_ineligible_roots.insert((node.height, node.hash.0));
             }
             let is_eligible_tip = store.view_is_eligible_header_tip(node.hash);
             let is_leaf = !store.view_header_has_children(node.hash);
@@ -146,10 +167,10 @@ impl RetentionCandidates {
                 match store.view_header_chain_score(node.hash) {
                     Ok(score) => {
                         if is_eligible_tip {
-                            eligible_tips.insert(score);
+                            eligible_header_tips.insert(score);
                         }
                         if is_leaf {
-                            leaves.insert(score);
+                            header_leaves.insert(score);
                         }
                     }
                     Err(graph_error) => error = Some(graph_error),
@@ -160,56 +181,56 @@ impl RetentionCandidates {
             return Err(error);
         }
         Ok(Self {
-            permanent_roots,
-            eligible_tips,
-            leaves,
-            scanned,
+            permanently_ineligible_roots,
+            eligible_header_tips,
+            header_leaves,
+            header_nodes_scanned,
         })
     }
 }
 
-fn protect_path<G: HeaderGraphView>(
+fn add_protected_header_path<G: HeaderGraphView>(
     store: &G,
-    tip: block::Hash,
-    protected: &mut HashSet<block::Hash>,
+    path_tip_hash: block::Hash,
+    protected_header_hashes: &mut HashSet<block::Hash>,
 ) -> Result<usize, GraphError> {
-    let mut hash = tip;
-    let mut visited = 0usize;
+    let mut header_hash = path_tip_hash;
+    let mut visited_header_nodes = 0usize;
     loop {
-        if protected.contains(&hash) {
-            return Ok(visited);
+        if protected_header_hashes.contains(&header_hash) {
+            return Ok(visited_header_nodes);
         }
-        visited = visited.saturating_add(1);
-        let node = store
-            .view_header_node(hash)
-            .ok_or(GraphError::UnknownHeaderNode(hash))?;
-        protected.insert(hash);
-        if hash == store.view_finalized_frontier().hash {
-            return Ok(visited);
+        visited_header_nodes = visited_header_nodes.saturating_add(1);
+        let header_node = store
+            .view_header_node(header_hash)
+            .ok_or(GraphError::UnknownHeaderNode(header_hash))?;
+        protected_header_hashes.insert(header_hash);
+        if header_hash == store.view_finalized_frontier().hash {
+            return Ok(visited_header_nodes);
         }
-        hash = node.parent_hash;
+        header_hash = header_node.parent_hash;
     }
 }
 
 fn evict_permanently_ineligible<G: HeaderGraphEdit>(
     store: &mut G,
-    protected: &HashSet<block::Hash>,
+    protected_header_hashes: &HashSet<block::Hash>,
     candidates: &mut RetentionCandidates,
-    work: &mut RetentionWork,
+    retention_work: &mut RetentionWork,
 ) -> Result<(), GraphError> {
-    while let Some((_, raw_hash)) = candidates.permanent_roots.pop_first() {
+    while let Some((_, raw_hash)) = candidates.permanently_ineligible_roots.pop_first() {
         let root = block::Hash(raw_hash);
-        if protected.contains(&root) {
+        if protected_header_hashes.contains(&root) {
             continue;
         }
         if store.view_header_node(root).is_none() {
             continue;
         }
-        work.graph_workspaces = work.graph_workspaces.saturating_add(1);
+        retention_work.graph_workspaces = retention_work.graph_workspaces.saturating_add(1);
         let mut descendants = subtree_postorder(store, root);
         for hash in descendants.drain(..) {
             store.edit_remove_header_leaf(hash)?;
-            work.evicted_nodes = work.evicted_nodes.saturating_add(1);
+            retention_work.evicted_nodes = retention_work.evicted_nodes.saturating_add(1);
         }
     }
     Ok(())
@@ -237,25 +258,27 @@ fn subtree_postorder<G: HeaderGraphView>(store: &G, root: block::Hash) -> Vec<bl
 
 fn evict_tip_branch<G: HeaderGraphEdit>(
     store: &mut G,
-    root: block::Hash,
-    protected: &HashSet<block::Hash>,
-    work: &mut RetentionWork,
+    branch_tip_hash: block::Hash,
+    protected_header_hashes: &HashSet<block::Hash>,
+    retention_work: &mut RetentionWork,
 ) -> Result<(), GraphError> {
-    if protected.contains(&root) || root == store.view_finalized_frontier().hash {
+    if protected_header_hashes.contains(&branch_tip_hash)
+        || branch_tip_hash == store.view_finalized_frontier().hash
+    {
         return Ok(());
     }
     let mut hash = store
-        .view_header_node(root)
-        .ok_or(GraphError::UnknownHeaderNode(root))?
+        .view_header_node(branch_tip_hash)
+        .ok_or(GraphError::UnknownHeaderNode(branch_tip_hash))?
         .parent_hash;
-    work.graph_workspaces = work.graph_workspaces.saturating_add(1);
-    for descendant in subtree_postorder(store, root) {
+    retention_work.graph_workspaces = retention_work.graph_workspaces.saturating_add(1);
+    for descendant in subtree_postorder(store, branch_tip_hash) {
         store.edit_remove_header_leaf(descendant)?;
-        work.evicted_nodes = work.evicted_nodes.saturating_add(1);
+        retention_work.evicted_nodes = retention_work.evicted_nodes.saturating_add(1);
     }
 
     loop {
-        if protected.contains(&hash) || hash == store.view_finalized_frontier().hash {
+        if protected_header_hashes.contains(&hash) || hash == store.view_finalized_frontier().hash {
             return Ok(());
         }
         let node = store
@@ -266,7 +289,7 @@ fn evict_tip_branch<G: HeaderGraphEdit>(
         }
         let parent = node.parent_hash;
         store.edit_remove_header_leaf(hash)?;
-        work.evicted_nodes = work.evicted_nodes.saturating_add(1);
+        retention_work.evicted_nodes = retention_work.evicted_nodes.saturating_add(1);
         if store.view_header_children(parent).is_empty() {
             hash = parent;
         } else {
@@ -276,517 +299,12 @@ fn evict_tip_branch<G: HeaderGraphEdit>(
 }
 
 #[cfg(feature = "test-support")]
-mod benchmark_support {
-    use std::{num::NonZeroUsize, sync::Arc};
-
-    use zakura_chain::block::genesis::regtest_genesis_block;
-
-    use super::*;
-    use crate::{HeaderValidationState, InsertResult, MemHeaderStore};
-
-    /// Reusable retained-chain fixture for release-mode retention benchmarks.
-    pub struct RetentionBenchmarkFixture {
-        graph: MemHeaderStore,
-        header_best: Frontier,
-        finalized: Frontier,
-        retained_non_finalized_nodes: usize,
-    }
-
-    impl RetentionBenchmarkFixture {
-        /// Build a linear graph at `percent` of the V1 retained-node limit.
-        pub fn at_v1_limit_percent(percent: usize) -> Result<Self, GraphError> {
-            assert!(matches!(percent, 25 | 50 | 90 | 100));
-            let block = regtest_genesis_block();
-            let hash = block.hash();
-            let work = block
-                .header
-                .difficulty_threshold
-                .to_work()
-                .expect("the regtest genesis target has canonical work");
-            let finalized = Frontier::new(block::Height(0), hash);
-            let mut graph =
-                MemHeaderStore::new(finalized, block.header.clone(), work, work.as_u256())?;
-            let retained_non_finalized_nodes =
-                crate::MAX_NON_FINALIZED_NODES_V1.saturating_mul(percent) / 100;
-            let mut header_best = finalized;
-            for index in 0..retained_non_finalized_nodes {
-                let mut header = *regtest_genesis_block().header;
-                header.previous_block_hash = header_best.hash;
-                header.nonce =
-                    [u8::try_from(index % 251).expect("the nonce marker fits"); 32].into();
-                header_best = match graph.insert(
-                    Arc::new(header),
-                    HeaderValidationState::Valid,
-                    [],
-                    BodyValidationState::Unknown,
-                )? {
-                    InsertResult::Inserted(frontier) | InsertResult::AlreadyPresent(frontier) => {
-                        frontier
-                    }
-                };
-            }
-            Ok(Self {
-                graph,
-                header_best,
-                finalized,
-                retained_non_finalized_nodes,
-            })
-        }
-
-        /// Measure the ordinary exact-limit check without graph traversal.
-        pub fn ordinary_check(&mut self) -> Result<RetentionBenchmarkResult, GraphError> {
-            self.enforce_with_limit(self.retained_non_finalized_nodes.max(1))
-        }
-
-        /// Measure refusal when the protected chain exceeds the node limit by one.
-        pub fn protected_refusal(&mut self) -> Result<RetentionBenchmarkResult, GraphError> {
-            self.enforce_with_limit(self.retained_non_finalized_nodes.saturating_sub(1).max(1))
-        }
-
-        fn enforce_with_limit(
-            &mut self,
-            node_limit: usize,
-        ) -> Result<RetentionBenchmarkResult, GraphError> {
-            let limits = EngineLimits {
-                max_non_finalized_nodes: NonZeroUsize::new(node_limit)
-                    .expect("the benchmark node limit is nonzero"),
-                ..EngineLimits::v1()
-            };
-            let plan = enforce_retention(
-                &mut self.graph,
-                self.header_best,
-                self.finalized,
-                [],
-                limits,
-            )?;
-            Ok(RetentionBenchmarkResult {
-                admission_refused: plan.admission_refused,
-                protected_path_visits: plan.work.protected_path_visits,
-                candidate_nodes_scanned: plan.work.candidate_nodes_scanned,
-                evicted_nodes: plan.work.evicted_nodes,
-                graph_workspaces: plan.work.graph_workspaces,
-            })
-        }
-    }
-
-    /// Structural result returned by release-mode retention benchmarks.
-    #[derive(Copy, Clone, Debug, Eq, PartialEq)]
-    pub struct RetentionBenchmarkResult {
-        /// Whether the protected chain refused admission.
-        pub admission_refused: bool,
-        /// Unique retained nodes visited while constructing the protected union.
-        pub protected_path_visits: usize,
-        /// Retained nodes scanned while constructing eviction candidates.
-        pub candidate_nodes_scanned: usize,
-        /// Retained nodes removed by eviction.
-        pub evicted_nodes: usize,
-        /// Graph-sized retention workspaces allocated during the attempt.
-        pub graph_workspaces: usize,
-    }
-}
+#[path = "retention/benchmark_support.rs"]
+mod benchmark_support;
 
 #[cfg(feature = "test-support")]
 pub use benchmark_support::{RetentionBenchmarkFixture, RetentionBenchmarkResult};
 
 #[cfg(test)]
-mod tests {
-    use std::{num::NonZeroUsize, sync::Arc};
-
-    use super::*;
-    use crate::{
-        BodyValidationState, EligibilityReason, HeaderValidationState, InsertResult, MemHeaderStore,
-    };
-    use zakura_chain::block::genesis::regtest_genesis_block;
-
-    fn store() -> MemHeaderStore {
-        let block = regtest_genesis_block();
-        let hash = block.hash();
-        let work = block
-            .header
-            .difficulty_threshold
-            .to_work()
-            .expect("valid work");
-        MemHeaderStore::new(
-            Frontier::new(block::Height(0), hash),
-            block.header.clone(),
-            work,
-            work.as_u256(),
-        )
-        .expect("fixture anchor matches")
-    }
-
-    fn insert(
-        store: &mut MemHeaderStore,
-        parent: block::Hash,
-        seed: u8,
-        reasons: impl IntoIterator<Item = EligibilityReason>,
-    ) -> Frontier {
-        let mut header = *regtest_genesis_block().header;
-        header.previous_block_hash = parent;
-        header.nonce = [seed; 32].into();
-        let header = Arc::new(header);
-        match store
-            .insert(
-                header,
-                HeaderValidationState::Valid,
-                reasons,
-                BodyValidationState::Unknown,
-            )
-            .expect("fixture parent is retained")
-        {
-            InsertResult::Inserted(frontier) | InsertResult::AlreadyPresent(frontier) => frontier,
-        }
-    }
-
-    fn limits(tips: usize, nodes: usize) -> EngineLimits {
-        EngineLimits {
-            max_candidate_tips: NonZeroUsize::new(tips).expect("test limit is nonzero"),
-            max_non_finalized_nodes: NonZeroUsize::new(nodes).expect("test limit is nonzero"),
-            ..EngineLimits::v1()
-        }
-    }
-
-    #[test]
-    fn candidate_tip_eviction_is_lowest_work_then_smallest_raw_hash() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let tips: Vec<_> = (1..=12)
-            .map(|seed| insert(&mut store, anchor.hash, seed, []))
-            .collect();
-        let header_best = store
-            .select_best_header_chain()
-            .expect("graph is coherent")
-            .0;
-        let mut expected: Vec<_> = tips
-            .iter()
-            .copied()
-            .filter(|tip| *tip != header_best)
-            .collect();
-        expected.sort_unstable_by_key(|tip| {
-            store
-                .header_chain_score(tip.hash)
-                .expect("retained")
-                .tip_hash
-                .0
-        });
-
-        let plan = enforce_retention(&mut store, header_best, anchor, [], limits(10, 100))
-            .expect("retention succeeds");
-        assert!(store.header_node(expected[0].hash).is_none());
-        assert!(store.header_node(expected[1].hash).is_none());
-        assert_eq!(store.eligible_header_tips().len(), 10);
-        assert!(store.header_node(header_best.hash).is_some());
-        assert!(!plan.admission_refused);
-        assert_eq!(plan.work.protected_path_visits, 2);
-        assert_eq!(plan.work.candidate_nodes_scanned, 13);
-        assert_eq!(plan.work.evicted_nodes, 2);
-        assert_eq!(plan.work.graph_workspaces, 4);
-
-        let reacquired_seed = (1..=12)
-            .find(|seed| {
-                let mut header = *regtest_genesis_block().header;
-                header.previous_block_hash = anchor.hash;
-                header.nonce = [*seed; 32].into();
-                header.hash() == expected[0].hash
-            })
-            .expect("the evicted fixture tip has a seed");
-        let reacquired = insert(&mut store, anchor.hash, reacquired_seed, []);
-        assert_eq!(reacquired.hash, expected[0].hash);
-        assert!(store
-            .header_node(reacquired.hash)
-            .expect("reacquired")
-            .is_eligible());
-    }
-
-    #[test]
-    fn eligible_tip_with_deferred_descendant_is_evicted_without_stalling() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let tips: Vec<_> = (1..=12)
-            .map(|seed| insert(&mut store, anchor.hash, seed, []))
-            .collect();
-        let header_best = store
-            .select_best_header_chain()
-            .expect("graph is coherent")
-            .0;
-        let victim = tips
-            .iter()
-            .copied()
-            .filter(|tip| *tip != header_best)
-            .min_by_key(|tip| store.header_chain_score(tip.hash).expect("retained"))
-            .expect("one unprotected tip exists");
-        let deferred = insert(&mut store, victim.hash, 13, []);
-        store
-            .set_header_validation_state(
-                deferred.hash,
-                HeaderValidationState::DeferredUntil(
-                    regtest_genesis_block().header.time + chrono::Duration::hours(3),
-                ),
-            )
-            .expect("the deferred descendant is retained");
-
-        assert!(store
-            .eligible_header_tips()
-            .iter()
-            .any(|tip| tip.hash == victim.hash));
-        assert!(!store
-            .header_node(deferred.hash)
-            .expect("the deferred child is retained")
-            .is_eligible());
-
-        let plan = enforce_retention(&mut store, header_best, anchor, [], limits(10, 100))
-            .expect("retention evicts the ineligible descendant subtree");
-
-        assert!(store.header_node(deferred.hash).is_none());
-        assert!(store.header_node(victim.hash).is_none());
-        assert_eq!(store.eligible_header_tips().len(), 10);
-        assert!(!plan.admission_refused);
-    }
-
-    #[test]
-    fn permanent_subtrees_are_evicted_first_only_under_pressure() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let permanent = insert(
-            &mut store,
-            anchor.hash,
-            1,
-            [EligibilityReason::CheckpointConflict {
-                height: block::Height(1),
-                expected: block::Hash([9; 32]),
-            }],
-        );
-        let selected = insert(&mut store, anchor.hash, 2, []);
-        let spare = insert(&mut store, anchor.hash, 3, []);
-
-        enforce_retention(&mut store, selected, anchor, [], limits(10, 2))
-            .expect("permanent subtree frees capacity");
-        assert!(store.header_node(permanent.hash).is_none());
-        assert!(store.header_node(selected.hash).is_some());
-        assert!(store.header_node(spare.hash).is_some());
-    }
-
-    #[test]
-    fn retention_below_both_limits_has_no_graph_work_or_effects() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let permanent = insert(
-            &mut store,
-            anchor.hash,
-            1,
-            [EligibilityReason::CheckpointConflict {
-                height: block::Height(1),
-                expected: block::Hash([9; 32]),
-            }],
-        );
-        let selected = insert(&mut store, anchor.hash, 2, []);
-        let before_header_node_count = store.header_node_count();
-        let before_tips = store.eligible_header_tips();
-        let before_permanent = store.header_node(permanent.hash).cloned();
-        let before_selected = store.header_node(selected.hash).cloned();
-
-        let plan = enforce_retention(&mut store, selected, anchor, [], limits(10, 10))
-            .expect("a graph below both limits needs no eviction planning");
-
-        assert_eq!(plan, RetentionPlan::default());
-        assert_eq!(plan.work, RetentionWork::default());
-        assert_eq!(store.header_node_count(), before_header_node_count);
-        assert_eq!(store.eligible_header_tips(), before_tips);
-        assert_eq!(store.header_node(permanent.hash), before_permanent.as_ref());
-        assert_eq!(store.header_node(selected.hash), before_selected.as_ref());
-    }
-
-    #[test]
-    fn protected_paths_and_context_references_fail_closed_under_node_pressure() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let first = insert(&mut store, anchor.hash, 1, []);
-        let selected = insert(&mut store, first.hash, 2, []);
-        let unprotected = insert(
-            &mut store,
-            anchor.hash,
-            3,
-            [EligibilityReason::CheckpointConflict {
-                height: block::Height(1),
-                expected: block::Hash([9; 32]),
-            }],
-        );
-
-        let plan = enforce_retention(&mut store, selected, anchor, [first.hash], limits(10, 1))
-            .expect("retention returns a typed refusal");
-        assert!(plan.admission_refused);
-        assert!(plan.resource_stalled);
-        assert!(store.header_node(selected.hash).is_some());
-        assert!(store.header_node(unprotected.hash).is_some());
-    }
-
-    #[test]
-    fn verified_side_node_does_not_become_a_protection_root() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let verified_body = insert(&mut store, anchor.hash, 1, []);
-        store
-            .set_body_validation_state(
-                verified_body.hash,
-                BodyValidationState::Verified {
-                    evidence: crate::EvidenceId::from_digest([0x51; 32]),
-                },
-            )
-            .expect("the full-state side branch becomes verified");
-        let selected_parent = insert(&mut store, anchor.hash, 2, []);
-        let selected = insert(&mut store, selected_parent.hash, 3, []);
-        let _header_only = insert(&mut store, anchor.hash, 4, []);
-
-        let plan = enforce_retention(&mut store, selected, selected, [], limits(1, 100))
-            .expect("retention evicts both unprotected side branches");
-
-        assert!(!plan.admission_refused);
-        assert!(store.header_node(selected.hash).is_some());
-        assert!(store.header_node(verified_body.hash).is_none());
-        assert_eq!(store.eligible_header_tips().len(), 1);
-    }
-
-    #[test]
-    fn unselected_verified_body_paths_do_not_fill_retention_capacity() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let first = insert(&mut store, anchor.hash, 1, []);
-        let verified_body = insert(&mut store, first.hash, 2, []);
-        store
-            .set_body_validation_state(
-                verified_body.hash,
-                BodyValidationState::Verified {
-                    evidence: crate::EvidenceId::from_digest([0x52; 32]),
-                },
-            )
-            .expect("the full-state branch becomes verified");
-        let selected = insert(&mut store, anchor.hash, 3, []);
-
-        let plan = enforce_retention(&mut store, selected, selected, [], limits(1, 1))
-            .expect("retention removes the unprotected verified side path");
-
-        assert!(!plan.admission_refused);
-        assert!(store.header_node(first.hash).is_none());
-        assert!(store.header_node(verified_body.hash).is_none());
-        assert!(store.header_node(selected.hash).is_some());
-    }
-
-    #[test]
-    fn authoritative_full_state_replacement_evicts_the_dropped_verified_branch() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let retained_tips = (1..=10)
-            .map(|seed| insert(&mut store, anchor.hash, seed, []))
-            .collect::<Vec<_>>();
-        for tip in &retained_tips {
-            store
-                .set_body_validation_state(
-                    tip.hash,
-                    BodyValidationState::Verified {
-                        evidence: crate::EvidenceId::from_digest(tip.hash.0),
-                    },
-                )
-                .expect("the full-state branch becomes verified");
-        }
-
-        let replacement = insert(&mut store, anchor.hash, 11, []);
-        store
-            .set_body_validation_state(
-                replacement.hash,
-                BodyValidationState::Verified {
-                    evidence: crate::EvidenceId::from_digest(replacement.hash.0),
-                },
-            )
-            .expect("the replacement full-state branch becomes verified");
-        let dropped = retained_tips[0];
-        let staged_tips = retained_tips
-            .iter()
-            .copied()
-            .skip(1)
-            .map(|tip| tip.hash)
-            .chain(std::iter::once(replacement.hash))
-            .collect::<Vec<_>>();
-
-        let plan = enforce_retention(
-            &mut store,
-            replacement,
-            replacement,
-            staged_tips,
-            limits(10, 100),
-        )
-        .expect("the authoritative full-state fork set permits replacement");
-
-        assert!(!plan.admission_refused);
-        assert!(store.header_node(dropped.hash).is_none());
-        assert!(store.header_node(replacement.hash).is_some());
-        assert_eq!(store.eligible_header_tips().len(), 10);
-    }
-
-    #[test]
-    fn overlapping_protected_paths_stop_at_their_first_protected_ancestor() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let first = insert(&mut store, anchor.hash, 1, []);
-        let selected = insert(&mut store, first.hash, 2, []);
-        let branch = insert(&mut store, first.hash, 3, []);
-        let mut protected = HashSet::new();
-
-        assert_eq!(
-            protect_path(&store, selected.hash, &mut protected)
-                .expect("the selected path is coherent"),
-            3
-        );
-        assert_eq!(
-            protect_path(&store, branch.hash, &mut protected)
-                .expect("the branch joins the selected path"),
-            1
-        );
-        assert_eq!(
-            protect_path(&store, first.hash, &mut protected)
-                .expect("the context reference is already protected"),
-            0
-        );
-        assert!(protected.contains(&anchor.hash));
-        assert!(protected.contains(&first.hash));
-        assert!(protected.contains(&selected.hash));
-        assert!(protected.contains(&branch.hash));
-    }
-
-    #[test]
-    fn reference_pruned_before_retention_is_no_longer_protectable() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-
-        let plan = enforce_retention(
-            &mut store,
-            anchor,
-            anchor,
-            [block::Hash([0x91; 32])],
-            limits(10, 100),
-        )
-        .expect("a reference already pruned by finality is ignored");
-
-        assert_eq!(plan, RetentionPlan::default());
-    }
-
-    #[test]
-    fn exact_v1_node_boundary_refuses_to_evict_the_selected_path() {
-        let mut store = store();
-        let anchor = store.finalized_frontier();
-        let mut selected = anchor;
-        for offset in 0..=crate::MAX_NON_FINALIZED_NODES_V1 {
-            let seed = u8::try_from(offset % 251).expect("the reduced test nonce fits in u8");
-            selected = insert(&mut store, selected.hash, seed, []);
-        }
-        assert_eq!(
-            store.header_node_count() - 1,
-            crate::MAX_NON_FINALIZED_NODES_V1 + 1
-        );
-
-        let plan = enforce_retention(&mut store, selected, anchor, [], EngineLimits::v1())
-            .expect("the exact boundary produces a typed refusal");
-        assert!(plan.admission_refused);
-        assert!(plan.resource_stalled);
-        assert!(store.header_node(selected.hash).is_some());
-    }
-}
+#[path = "retention/tests.rs"]
+mod tests;
